design : 사이드바 스타일링
